refactor: cleanup slideshow template
